Refactor (move test_update_at)
